Img2ImgExecutor/requirements.txt
diffusers
jina>=3.7
transformers>=4.21
torchvision
scipy
gdown